计算每一排的分割线

+++ core/src/kw/mulitplay/game/constant/Constant.java
@@ -20,5 +20,5 @@ public class Constant {
 
     public static float speed = 0;
     public static Instrument instrument;
-    public static int bpm;
+    public static float bpm;
 }
